+++ GAAF/inference.py
@@ -1,4 +1,4 @@
-from AutoGAF.inference_module import Locator_inference_module, setup_argparse
+from GAAF.inference_module import Locator_inference_module, setup_argparse
 
 args = setup_argparse()
 inference_module = Locator_inference_module(args)
