+++ GAAF/preprocess.py
@@ -1,4 +1,4 @@
-from AutoGAF.preprocessing_module import Preprocessor, setup_argparse
+from GAAF.preprocessing_module import Preprocessor, setup_argparse
 
 args = setup_argparse()
 preprocessor = Preprocessor(args)
