+++ GAAF/train.py
@@ -1,4 +1,4 @@
-from AutoGAF.training_module import train_locator, setup_argparse
+from GAAF.training_module import train_locator, setup_argparse
 
 args = setup_argparse()
 train_locator = train_locator(args)
